+++ vllm_ascend/patch/worker/__init__.py
@@ -39,9 +39,9 @@ import vllm_ascend.patch.worker.patch_gdn_attn  # noqa
 if is_310p():
     import vllm_ascend.patch.worker.patch_qwen3_5_310  # noqa
 else:
-    import vllm_ascend.patch.worker.patch_qwen3_next  # noqa
     import vllm_ascend.patch.worker.patch_qwen3_next_mtp  # noqa
     import vllm_ascend.patch.worker.patch_qwen3_5  # noqa
+import vllm_ascend.patch.worker.patch_qwen3_5  # noqa
 import vllm_ascend.patch.worker.patch_rejection_sampler  # noqa
 import vllm_ascend.patch.worker.patch_v2.patch_eagle  # noqa
 import vllm_ascend.patch.worker.patch_v2.patch_uva  # noqa
